chore: BeforeStudyTemplate 컴포넌트명 변경
